DSP-task4/requirements.txt
matplotlib==3.8.0
numpy==1.26.4
opencv_python==4.8.1.78
Pillow==10.1.0
Pillow==10.2.0
PyQt5==5.15.10
PyQt5_sip==12.12.2
pyqtgraph==0.13.3
